Refactoring: - Order class moved to a separated file - Worker class moved to a separated file - Unpack order function moved to a separated file
