+++ binding.gyp
@@ -5,7 +5,7 @@
         "<!(node -e \"require('nan')\")"
       ],
       "target_name": "async",
-      "sources": [ "async.cc" ]
+      "sources": [ "async.cc", "order.cc", "worker.cc", "order_helper.cc" ]
     }
   ]
 }
