+++ order.cc
@@ -1,0 +1,3 @@
+#include "order.h"
+
+Order::Order(int t): total(t) {}
