+++ order.h
@@ -1,0 +1,10 @@
+#ifndef ORDER_H
+#define ORDER_H
+
+class Order {
+  public:
+    Order(int t);
+    int total;
+};
+
+#endif
