+++ order_helper.h
@@ -1,0 +1,9 @@
+#ifndef ORDER_HELPER_H
+#define ORDER_HELPER_H
+
+#include "order.h"
+#include <v8.h>
+
+Order unpack_order(v8::Local<v8::Object> order_obj);
+
+#endif
